+++ frontend/src/components/common/common.ts
@@ -7,7 +7,6 @@ export { default as Header } from './header/header';
 export { default as Checkbox } from './checkbox/checkbox';
 export { default as NotFound } from './not-found/not-found';
 export { default as PrivateRoute } from './private-route/private-route';
-export { default as Episode } from './episode/episode';
 export { default as AuthWrapper } from './auth-wrapper/auth-wrapper';
 export { default as AuthPrivateRouter } from './auth-private-router/auth-private-router';
 export { default as AuthPublicRouter } from './auth-public-router/auth-public-router';
