+++ frontend/src/components/episode/styles.module.scss
@@ -17,14 +17,12 @@
     .description{
       flex: .6;
       h3{
-        margin-top: 10px;
+        margin: 10px 0 30px 0;
         font-size: 40px;
         font-weight: bold;
       }
-      h5{
+      h1{
         margin: 0;
-        font-size: 18px;
-        font-weight: medium;
         color: #ff7cc2;
       }
       p{
@@ -34,8 +32,6 @@
     .logo{
       flex: .4;
       img{
-        width: 280px;
-        height: 280px;
         border: 2px solid #e9e9e9;
         margin: auto;
       }
