+++ frontend/src/store/actions.ts
@@ -1,3 +1,3 @@
 export * as auth from './auth/actions';
 export * as homepage from './homepage/actions';
-export * as episodepage from './episodepage/actions';
+export * as episode from './episode/actions';
