+++ frontend/src/store/episode/common.ts
@@ -1,6 +1,5 @@
 enum ActionType {
   LOAD_EPISODE = 'episodes/load-episode',
-  LOAD_EPISODES = 'episodes/load-episodes',
 }
 
 export { ActionType };
