+++ frontend/src/store/episode/reducer.ts
@@ -1,18 +1,16 @@
 import { createReducer } from '@reduxjs/toolkit';
 import { DataStatus } from 'common/enums/enums';
 import { Episode } from 'common/types/types';
-import { loadEpisode, loadEpisodes } from './actions';
+import { loadEpisode } from './actions';
 
 type State = {
   dataStatus: DataStatus;
   episode: Episode | null;
-  episodes: Episode[];
 };
 
 const initialState: State = {
   dataStatus: DataStatus.IDLE,
   episode: null,
-  episodes: [],
 };
 
 const reducer = createReducer(initialState, (builder) => {
@@ -24,16 +22,6 @@ const reducer = createReducer(initialState, (builder) => {
     state.episode = action.payload;
   });
   builder.addCase(loadEpisode.rejected, (state) => {
-    state.dataStatus = DataStatus.REJECTED;
-  });
-  builder.addCase(loadEpisodes.pending, (state) => {
-    state.dataStatus = DataStatus.PENDING;
-  });
-  builder.addCase(loadEpisodes.fulfilled, (state, action) => {
-    state.dataStatus = DataStatus.FULFILLED;
-    state.episodes = action.payload;
-  });
-  builder.addCase(loadEpisodes.rejected, (state) => {
     state.dataStatus = DataStatus.REJECTED;
   });
 });
