+++ frontend/src/store/root-reducer.ts
@@ -1,12 +1,12 @@
 import { reducer as auth } from './auth/reducer';
 import { reducer as homepage } from './homepage/reducer';
 import { reducer as toastrReducer } from 'react-redux-toastr';
-import { reducer as episodepage } from './episodepage/reducer';
+import { reducer as episode } from './episode/reducer';
 
 const rootReducer = {
   auth,
   homepage,
-  episodepage,
+  episode,
   toastr: toastrReducer,
 };
